make sure fd_table->offset is always a 32bit integer during memcpy() operations in memdev.c

+++ drivers/char/memdev.c
@@ -382,7 +382,7 @@ int mem_read(struct inode *i, struct fd *fd_table, char *buffer, __size_t count)
 		return 0;
 	}
 	count = MIN(count, physical_memory - fd_table->offset);
-	memcpy_b(buffer, (void *)P2V(fd_table->offset), count);
+	memcpy_b(buffer, (void *)P2V((unsigned int)fd_table->offset), count);
 	fd_table->offset += count;
 	return count;
 }
@@ -396,7 +396,7 @@ int mem_write(struct inode *i, struct fd *fd_table, const char *buffer, __size_t
 		return 0;
 	}
 	count = MIN(count, physical_memory - fd_table->offset);
-	memcpy_b((void *)P2V(fd_table->offset), buffer, count);
+	memcpy_b((void *)P2V((unsigned int)fd_table->offset), buffer, count);
 	fd_table->offset += count;
 	return count;
 }
@@ -422,7 +422,7 @@ int kmem_read(struct inode *i, struct fd *fd_table, char *buffer, __size_t count
 
 	physical_memory = P2V((kstat.physical_pages << PAGE_SHIFT));
 	if(P2V(fd_table->offset + count) < physical_memory) {
-		memcpy_b(buffer, (void *)P2V(fd_table->offset), count);
+		memcpy_b(buffer, (void *)P2V((unsigned int)fd_table->offset), count);
 		fd_table->offset += count;
 	} else {
 		count = 0;
@@ -436,7 +436,7 @@ int kmem_write(struct inode *i, struct fd *fd_table, const char *buffer, __size_
 
 	physical_memory = P2V((kstat.physical_pages << PAGE_SHIFT));
 	if(P2V(fd_table->offset + count) < physical_memory) {
-		memcpy_b((void *)P2V(fd_table->offset), buffer, count);
+		memcpy_b((void *)P2V((unsigned int)fd_table->offset), buffer, count);
 		fd_table->offset += count;
 	} else {
 		count = 0;
